requirements.txt
Django
whitenoise
gunicorn==19.6.0
dj-database-url
djangorestframework
slackclient
PyJWT
schedule
mixpanel==4.0.0